requirements.txt
Flask==2.3.3 
gunicorn==21.2.0 
google-cloud-texttospeech==2.17.2 
python-dotenv==1.0.0 
flask-cors==4.0.1